fix: promote waitlist when max_attendees increases via event edit
Editing an event's max_attendees upward via PATCH updated the DB
column but never re-ran waitlist promotion, so newly-opened spots
stayed empty and the displayed attending count read low until a
later RSVP change happened to trigger promotion.

Fixes Issue 1264.
